Organize tests and deduplicate mocked resources
